Don't try to fill up the entire buffer on a read.
An LZ4 frame is designed so that partial blocks can be "flushed" by
the encoder at any time.  A frame can be a long term stream sent over
a network connection.  The decoder needs to be sure that it only calls
the underlying read once at a time, when necessary, and be able to
return partially filled buffers.  This allows the caller to respond
to flushed data without the decoder blocking on a network read.