Fix pointer invalidation on Decoder move
When the Decoder is moved, so is Decoder::buf and its address. This
causes pointer invalidation in DecoderContext, leading to errors and
possible segmentation faults. Therefore, Decoder::buf is set to a fixed
address, in exchange for an extra one-shot memory allocation on the
heap. To enforce the constness of the buffer length, a boxed slice is
used.