Allow delegating all linking
This is helpful when an external build system (e.g. Buck) is taking care
of linking liblz4, and we don't want the Rust compiler to emit any
linking arguments.

To do so, this adds a flag to just bail out of the build script and not
build any native crate if LZ4_SYS_DELEGATE_LINKING is set in the
environment.